requirements.txt
cachetools==5.2.0 ; python_version >= "3.11" and python_version < "4.0"
certifi==2022.9.24 ; python_version >= "3.11" and python_version < "4"
charset-normalizer==2.1.1 ; python_version >= "3.11" and python_version < "4"
croniter==1.3.7 ; python_version >= "3.11" and python_version < "4.0"
google-api-core==2.10.2 ; python_version >= "3.11" and python_version < "4.0"
google-auth==2.14.1 ; python_version >= "3.11" and python_version < "4.0"
google-cloud-core==2.3.2 ; python_version >= "3.11" and python_version < "4.0"
google-cloud-storage==2.6.0 ; python_version >= "3.11" and python_version < "4.0"
google-crc32c==1.5.0 ; python_version >= "3.11" and python_version < "4.0"
google-resumable-media==2.4.0 ; python_version >= "3.11" and python_version < "4.0"
googleapis-common-protos==1.57.0 ; python_version >= "3.11" and python_version < "4.0"
idna==3.4 ; python_version >= "3.11" and python_version < "4"
protobuf==4.21.9 ; python_version >= "3.11" and python_version < "4.0"
pyasn1-modules==0.2.8 ; python_version >= "3.11" and python_version < "4.0"
pyasn1==0.4.8 ; python_version >= "3.11" and python_version < "4.0"
python-dateutil==2.8.2 ; python_version >= "3.11" and python_version < "4.0"
requests==2.28.1 ; python_version >= "3.11" and python_version < "4"
rsa==4.9 ; python_version >= "3.11" and python_version < "4"
six==1.16.0 ; python_version >= "3.11" and python_version < "4.0"
urllib3==1.26.12 ; python_version >= "3.11" and python_version < "4"
